+++ src/reconnectcontroller/ReconnectController.ts
@@ -19,11 +19,6 @@ export default interface ReconnectController {
   startedConnectionAttempt(isFirstConnection: boolean): void;
 
   /**
-   * Returns the count of retries for the session
-   */
-  getRetryCount(): number;
-
-  /**
    * Indicates whether a connection attempt is already in progress.
    */
   hasStartedConnectionAttempt(): boolean;