+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - [Demo] Add default SSE to meeting notifications queue in CF template
+- Add meeting events
 
 ### Changed
 - Removed check for iceGatheringState to be complete for bypassing gathering ice candidate again